Ora tutti gli endpoint sono pronti per l'autenticazione

+++ src/main/java/me/fagiolini/cinemapp/controller/CinemaController.java
@@ -1,10 +1,13 @@
 package me.fagiolini.cinemapp.controller;
 
+import io.micronaut.http.HttpRequest;
 import io.micronaut.http.annotation.*;
 import io.micronaut.security.annotation.Secured;
 import io.micronaut.security.rules.SecurityRule;
 import jakarta.inject.Inject;
+import me.fagiolini.cinemapp.auth.VerifyAdmin;
 import me.fagiolini.cinemapp.db.tables.pojos.Cinema;
+import me.fagiolini.cinemapp.exception.myException;
 import me.fagiolini.cinemapp.service.CinemaService;
 
 import java.util.List;
@@ -23,18 +26,23 @@ public class CinemaController {
     public Cinema getCinemaById(@PathVariable Integer id) {
         return this.cinemaService.getCinemaById(id);
     }
+    @Secured(SecurityRule.IS_AUTHENTICATED)
     @Post(uri = "/insertCinema")
-    public void insertCinema(@Body Cinema cinema) {
-        this.cinemaService.saveCinema(cinema);
+    public void insertCinema(@Body Cinema cinema, HttpRequest<?> request) throws myException {
+        if(VerifyAdmin.verify(request))
+            this.cinemaService.saveCinema(cinema);
     }
+    @Secured(SecurityRule.IS_AUTHENTICATED)
     @Put(uri = "/updateCinema")
-    public void updateCinema(@Body Cinema cinema) {
-        this.cinemaService.updateCinema(cinema);
+    public void updateCinema(@Body Cinema cinema, HttpRequest<?> request) throws myException {
+        if(VerifyAdmin.verify(request))
+            this.cinemaService.updateCinema(cinema);
     }
-
+    @Secured(SecurityRule.IS_AUTHENTICATED)
     @Delete(uri = "/deleteCinema/{id}")
-    public void deleteCinema(@PathVariable long id) {
-        this.cinemaService.deleteCinema(id);
+    public void deleteCinema(@PathVariable long id, HttpRequest<?> request) throws myException {
+        if(VerifyAdmin.verify(request))
+            this.cinemaService.deleteCinema(id);
     }
