+++ src/main/java/me/fagiolini/cinemapp/controller/FilmController.java
@@ -1,10 +1,13 @@
 package me.fagiolini.cinemapp.controller;
 
+import io.micronaut.http.HttpRequest;
 import io.micronaut.http.annotation.*;
 import io.micronaut.security.annotation.Secured;
 import io.micronaut.security.rules.SecurityRule;
 import jakarta.inject.Inject;
+import me.fagiolini.cinemapp.auth.VerifyAdmin;
 import me.fagiolini.cinemapp.db.tables.pojos.Film;
+import me.fagiolini.cinemapp.exception.myException;
 import me.fagiolini.cinemapp.repository.FilmRepository;
 import me.fagiolini.cinemapp.service.FilmService;
 
@@ -25,19 +28,22 @@ public class FilmController {
     public Film getFilm(@PathVariable Long id) {
         return this.filmService.getFilmById(id);
     }
-
+    @Secured(SecurityRule.IS_AUTHENTICATED)
     @Post(uri = "/insertFilm")
-    public void insertFilm(@Body Film film) {
+    public void insertFilm(@Body Film film, HttpRequest<?> request) throws myException {
+        if(VerifyAdmin.verify(request))
         this.filmService.insertFilm(film);
     }
-
+    @Secured(SecurityRule.IS_AUTHENTICATED)
     @Put(uri = "/updateFilm")
-    public void updateFilm(@Body Film film) {
+    public void updateFilm(@Body Film film, HttpRequest<?> request) throws myException {
+        if(VerifyAdmin.verify(request))
         this.filmService.updateFilm(film);
     }
-
+    @Secured(SecurityRule.IS_AUTHENTICATED)
     @Delete(uri = "/deleteFilm/{id}")
-    public void deleteFilm(@PathVariable Long id) {
+    public void deleteFilm(@PathVariable Long id, HttpRequest<?> request) throws myException {
+        if(VerifyAdmin.verify(request))
         this.filmService.deleteFilmById(id);
     }
     @Secured(SecurityRule.IS_ANONYMOUS)
